+++ src/lib/components/table/proposals/ResolvedHeight.tsx
@@ -1,0 +1,25 @@
+import { Text } from "@chakra-ui/react";
+
+import { ExplorerLink } from "lib/components/ExplorerLink";
+
+import type { ProposalsTableRowProps } from "./ProposalsTableRow";
+
+export const ResolvedHeight = ({
+  resolvedHeight,
+  isInactive,
+  isDepositOrVoting,
+}: {
+  resolvedHeight: ProposalsTableRowProps["proposal"]["resolvedHeight"];
+  isInactive: boolean;
+  isDepositOrVoting: boolean;
+}) => {
+  if (isDepositOrVoting) return <Text color="text.dark">Pending</Text>;
+  if (!resolvedHeight || isInactive) return <Text color="text.dark">N/A</Text>;
+  return (
+    <ExplorerLink
+      type="block_height"
+      value={resolvedHeight.toString()}
+      showCopyOnHover
+    />
+  );
+};